Remove Messenger classes and use PeerClient instead
