+++ src/Catalyst.Core.Modules.Ledger.Tests/UnitTests/LedgerTests.cs
@@ -86,7 +86,7 @@ namespace Catalyst.Core.Modules.Ledger.Tests.UnitTests
         {
             _testScheduler = new TestScheduler();
             _fakeRepository = Substitute.For<IAccountRepository>();
-            _hashProvider = new HashProvider(HashingAlgorithm.GetAlgorithmMetadata("blake2b-256"));
+            _hashProvider = new HashProvider(HashingAlgorithm.GetAlgorithmMetadata("keccak-256"));
             _mapperProvider = new TestMapperProvider();
 
             _logger = Substitute.For<ILogger>();
