--- src/Catalyst.Core.Modules.Sync/Manager/Messenger.cs
@@ -1,55 +1,0 @@
-#region LICENSE
-
-/**
-* Copyright (c) 2019 Catalyst Network
-*
-* This file is part of Catalyst.Node <https://github.com/catalyst-network/Catalyst.Node>
-*
-* Catalyst.Node is free software: you can redistribute it and/or modify
-* it under the terms of the GNU General Public License as published by
-* the Free Software Foundation, either version 2 of the License, or
-* (at your option) any later version.
-*
-* Catalyst.Node is distributed in the hope that it will be useful,
-* but WITHOUT ANY WARRANTY; without even the implied warranty of
-* MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-* GNU General Public License for more details.
-*
-* You should have received a copy of the GNU General Public License
-* along with Catalyst.Node. If not, see <https://www.gnu.org/licenses/>.
-*/
-
-#endregion
-
-using Catalyst.Abstractions.P2P;
-using Catalyst.Core.Lib.Extensions;
-using Catalyst.Core.Lib.IO.Messaging.Dto;
-using Catalyst.Protocol.Peer;
-using Google.Protobuf;
-using System.Collections.Generic;
-
-namespace Catalyst.Core.Modules.Sync.Manager
-{
-    public class Messenger : IMessenger
-    {
-        private IPeerClient _peerClient;
-        private IPeerSettings _peerSettings;
-
-        public Messenger(IPeerClient peerClient, IPeerSettings peerSettings)
-        {
-            _peerClient = peerClient;
-            _peerSettings = peerSettings;
-        }
-
-        public void SendMessageToPeers(IMessage message, IEnumerable<PeerId> peers)
-        {
-            var protocolMessage = message.ToProtocolMessage(_peerSettings.PeerId);
-            foreach (var peer in peers)
-            {
-                _peerClient.SendMessage(new MessageDto(
-                    protocolMessage,
-                    peer));
-            }
-        }
-    }
-}
