+++ src/Catalyst.Core.Modules.Sync/Synchroniser.cs
@@ -204,13 +204,8 @@ namespace Catalyst.Core.Modules.Sync
                     {
                         if(DeltaCache.TryGetOrAddConfirmedDelta(deltaIndex.Cid, out Delta _))
                         {
-                            return;
+                            continue;
                         }
-                        //var deltaStream =
-                        //    await _dfsService.UnixFsApi.ReadFileAsync(deltaIndex.Cid).ConfigureAwait(false);
-                        //await _dfsService.UnixFsApi
-                        //   .AddAsync(deltaStream, options: new AddFileOptions { Hash = _hashProvider.HashingAlgorithm.Name })
-                        //   .ConfigureAwait(false);
                     }
                     catch (Exception exc) { }
